created functionality with login and signup buttons. split CSS and React files and fixed pathing
