+++ client/src/components/Header/Header.jsx
@@ -18,6 +18,7 @@ import {
 } from "@chakra-ui/react";
 
 import { HamburgerIcon, CloseIcon, AddIcon } from "@chakra-ui/icons";
+import { Link } from "react-router-dom";
 
 const Links = ["Culinary Kickstart"];
 
@@ -67,36 +68,33 @@ export default function HeaderBar() {
             </HStack>
           </HStack>
           <Flex alignItems={"center"}>
-            <Button
-              variant={"solid"}
-              colorScheme={"teal"}
-              size={"sm"}
-              mr={4}
-              leftIcon={<AddIcon />}
-            >
-              Action
-            </Button>
+            <Link to={"/login"}>
+              <Button variant={"solid"} colorScheme={"teal"} size={"sm"} mr={4}>
+                Login
+              </Button>
+            </Link>
+            <Link to={"/signup"}>
+              <Button variant={"solid"} colorScheme={"teal"} size={"sm"} mr={4}>
+                Sign Up
+              </Button>
+            </Link>
             <Menu>
-              <MenuButton
-                as={Button}
-                rounded={"full"}
-                variant={"link"}
-                cursor={"pointer"}
-                minW={0}
-              >
-                <Avatar
-                  size={"sm"}
-                  src={
-                    "https://images.unsplash.com/photo-1493666438817-866a91353ca9?ixlib=rb-0.3.5&q=80&fm=jpg&crop=faces&fit=crop&h=200&w=200&s=b616b2c5b373a80ffc9636ba24f7a4a9"
-                  }
-                />
-              </MenuButton>
-              <MenuList>
-                <MenuItem>Link 1</MenuItem>
-                <MenuItem>Link 2</MenuItem>
-                <MenuDivider />
-                <MenuItem>Link 3</MenuItem>
-              </MenuList>
+              <Link to={"/profile"}>
+                <MenuButton
+                  as={Button}
+                  rounded={"full"}
+                  variant={"link"}
+                  cursor={"pointer"}
+                  minW={0}
+                >
+                  <Avatar
+                    size={"sm"}
+                    src={
+                      "https://images.unsplash.com/photo-1493666438817-866a91353ca9?ixlib=rb-0.3.5&q=80&fm=jpg&crop=faces&fit=crop&h=200&w=200&s=b616b2c5b373a80ffc9636ba24f7a4a9"
+                    }
+                  />
+                </MenuButton>
+              </Link>
             </Menu>
           </Flex>
         </Flex>
