+++ client/src/pages/Profile/Profile.css
@@ -1,0 +1,23 @@
+.main {
+  display: flex;
+  flex-direction: column;
+  align-items: stretch;
+  height: 100vh;
+}
+
+.mainFlexContainer {
+  display: inline-flex;
+  justify-content: flex-start;
+  align-content: flex-start;
+}
+
+.sideBarNav {
+  display: inline-flex;
+  flex-direction: column;
+  position: relative;
+  width: 7vw;
+}
+
+.searchContainer {
+  width: 100vw;
+}
